refactor(sdk): move encoding logic into separate module
